refactor(user.model): rename user controller
