+++ src/__tests__/ticket.test.ts
@@ -3,7 +3,7 @@ import request from 'supertest';
 
 import * as db from './configs/setup';
 import app from '../app';
-import { generateAccessToken } from '../controllers/user.controllers';
+import { generateAccessToken } from '../controllers/user.controller';
 import UserModel from '../models/user.model';
 
 describe('Tickets routes', () => {
